Rename index.ts to completionUtils.ts in Loki monaco completion provider
This is a follow up to #76374. The name `index.ts` is not matching with
the pattern in the whole project as per the original discussion. It was
noted in https://github.com/grafana/grafana/pull/76080#discussion_r1352812657
