+++ public/app/plugins/datasource/loki/components/monaco-query-field/monaco-completion-provider/completionUtils.ts
