+++ public/app/plugins/datasource/loki/components/monaco-query-field/monaco-completion-provider/completions.test.ts
@@ -6,10 +6,9 @@ import { LokiDatasource } from '../../../datasource';
 import { createLokiDatasource } from '../../../mocks';
 
 import { CompletionDataProvider } from './CompletionDataProvider';
+import { calculateRange } from './completionUtils';
 import { getAfterSelectorCompletions, getCompletions } from './completions';
 import { getSituation, Label, Situation } from './situation';
-
-import { calculateRange } from './index';
 
 jest.mock('../../../querybuilder/operations', () => ({
   explainOperator: () => 'Operator docs',
